refactor optimizer code
